# site/static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Promptfoo Local Server API",
    "version": "latest",
    "description": "OpenAPI document generated from the shared Zod DTO schemas used by the Promptfoo local server and web UI."
  },
  "servers": [
    {
      "url": "http://localhost:15500",
      "description": "Default local Promptfoo server"
    }
  ],
...
  "paths": {
...
    "/api/providers/test": {
      "post": {
        "operationId": "testProvider",
        "tags": ["Providers"],
        "summary": "Test a provider configuration",
        "requestBody": {
          "description": "JSON request body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string"
                  },
                  "providerOptions": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": ["id"],
                    "additionalProperties": {}
                  }
                },
                "required": ["providerOptions"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "testResult": {
                      "type": "object",
                      "properties": {
                        "success": {
                          "type": "boolean"
                        },
                        "message": {
                          "type": "string"
                        },
                        "error": {
                          "type": "string"
                        },
                        "changes_needed": {
                          "type": "boolean"
                        },
                        "changes_needed_reason": {
                          "type": "string"
                        },
                        "changes_needed_suggestions": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      },
                      "required": ["success", "message"],
                      "additionalProperties": {}
                    },
                    "providerResponse": {},
                    "transformedRequest": {}
                  },
                  "required": ["testResult"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "403": {
            "description": "CSRF protection rejected request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/api/providers/test-session": {
      "post": {
        "operationId": "testProviderSession",
        "tags": ["Providers"],
        "summary": "Test multi-turn provider session behavior",
        "requestBody": {
          "description": "JSON request body",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "provider": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "minLength": 1
                      }
                    },
                    "required": ["id"],
                    "additionalProperties": {}
                  },
                  "sessionConfig": {
                    "type": "object",
                    "properties": {
                      "sessionSource": {
                        "type": "string"
                      },
                      "sessionParser": {
                        "type": "string"
                      }
                    }
                  },
                  "mainInputVariable": {
                    "type": "string"
                  }
                },
                "required": ["provider"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "reason": {
                      "type": "string"
                    },
                    "error": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "properties": {
                        "sessionId": {
                          "type": "string"
                        },
                        "sessionSource": {
                          "type": "string"
                        },
                        "request1": {
                          "type": "object",
                          "properties": {
                            "prompt": {
                              "type": "string"
                            },
                            "sessionId": {
                              "type": "string"
                            }
                          },
                          "required": ["prompt"]
                        },
                        "response1": {},
                        "request2": {
                          "type": "object",
                          "properties": {
                            "prompt": {
                              "type": "string"
                            },
                            "sessionId": {
                              "type": "string"
                            }
                          },
                          "required": ["prompt"]
                        },
                        "response2": {},
                        "hasSessionIdTemplate": {
                          "type": "boolean"
                        },
                        "hasSessionParser": {
                          "type": "boolean"
                        },
                        "sessionParser": {
                          "type": "string"
                        }
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": ["success", "message"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "403": {
            "description": "CSRF protection rejected request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {},
                    "suggestion": {
                      "type": "string"
                    },
                    "success": {
                      "type": "boolean",
                      "enum": [false]
                    }
                  },
                  "required": ["error"],
                  "additionalProperties": {}
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}