# src/server/routes/providers.ts
import { Router } from 'express';
import { getEnvString } from '../../envars';
import { cloudConfig } from '../../globalConfig/cloud';
import logger from '../../logger';
import { testProviderConnectivity, testProviderSession } from '../../node/testProvider';
import { createTransformRequest, createTransformResponse } from '../../providers/httpTransforms';
import { loadApiProvider } from '../../providers/index';
import {
  doTargetPurposeDiscovery,
  type TargetPurposeDiscoveryResult,
} from '../../redteam/commands/discover';
import { neverGenerateRemote } from '../../redteam/remoteGeneration';
import { ProviderSchemas } from '../../types/api/providers';
import { ApiRoutes } from '../../types/api/routes';
import { fetchWithProxy } from '../../util/fetch/index';
import { getAvailableProviders } from '../config/serverConfig';
import { replyError, replyValidationError, sendError } from '../utils/errors';
import type { Request, Response } from 'express';

import type { ProviderOptions } from '../../types/providers';

export const providersRouter = Router();

/**
 * GET /api/providers/config-status
 *
 * Returns whether a custom provider configuration exists.
 * Used by redteam setup UI to determine whether to filter provider types.
 *
 * When custom config exists (hasCustomConfig: true), redteam setup restricts
 * provider types to: http, websocket, python, javascript for testing custom implementations.
 *
 * Response:
 * - hasCustomConfig: Boolean indicating if ui-providers.yaml exists with providers
 */
providersRouter.get(
  ApiRoutes.Providers.ConfigStatus.routerPath,
  (_req: Request, res: Response): void => {
    try {
      const serverProviders = getAvailableProviders();
      const hasCustomConfig = serverProviders.length > 0;

      res.json(
        ProviderSchemas.ConfigStatus.Response.parse({ success: true, data: { hasCustomConfig } }),
      );
    } catch (error) {
      sendError(res, 500, 'Failed to load provider config status', error);
    }
  },
);

providersRouter.post(
  ApiRoutes.Providers.Test.routerPath,
  async (req: Request, res: Response): Promise<void> => {
    const bodyResult = ProviderSchemas.Test.Request.safeParse(req.body);
    if (!bodyResult.success) {
      replyValidationError(res, bodyResult.error);
      return;
    }

    const { providerOptions } = bodyResult.data;

    try {
      const loadedProvider = await loadApiProvider(providerOptions.id, {
        options: {
          ...(providerOptions as ProviderOptions),
          config: {
            ...providerOptions.config,
            maxRetries: 1,
          },
        },
      });

      // Pass inputs explicitly from providerOptions since loaded provider may not expose config.inputs
      // Check both top-level inputs (from redteam UI) and config.inputs for backwards compatibility
      const result = await testProviderConnectivity({
        provider: loadedProvider,
        prompt: bodyResult.data.prompt,
        inputs: providerOptions.inputs || providerOptions.config?.inputs,
      });

      res.status(200).json(
        ProviderSchemas.Test.Response.parse({
          testResult: {
            success: result.success,
            message: result.message,
            error: result.error,
            changes_needed: result.analysis?.changes_needed,
            changes_needed_reason: result.analysis?.changes_needed_reason,
            changes_needed_suggestions: result.analysis?.changes_needed_suggestions,
          },
          providerResponse: result.providerResponse,
          transformedRequest: result.transformedRequest,
        }),
      );
    } catch (error) {
      sendError(res, 500, 'Failed to test provider', error);
    }
  },
);

providersRouter.post(
  ApiRoutes.Providers.Discover.routerPath,
  async (
    req: Request,
    res: Response<TargetPurposeDiscoveryResult | { error: string }>,
  ): Promise<void> => {
    const bodyResult = ProviderSchemas.Discover.Request.safeParse(req.body);
    if (!bodyResult.success) {
      replyValidationError(res, bodyResult.error);
      return;
    }
    const providerOptions = bodyResult.data;

    // Check that remote generation is enabled:
    if (neverGenerateRemote()) {
      replyError(res, 400, 'Requires remote generation be enabled.');
      return;
    }

    try {
      const loadedProvider = await loadApiProvider(providerOptions.id, {
        options: providerOptions as ProviderOptions,
      });
      const result = await doTargetPurposeDiscovery(loadedProvider, undefined, false);

      if (result) {
        res.json(ProviderSchemas.Discover.Response.parse(result));
      } else {
        replyError(res, 500, "Discovery failed to discover the target's purpose.");
      }
    } catch (e) {
      logger.error('Error calling target purpose discovery', {
        error: e,
        providerOptions,
      });
      sendError(res, 500, "Discovery failed to discover the target's purpose");
      return;
    }
  },
);

providersRouter.post(
  ApiRoutes.Providers.HttpGenerator.routerPath,
  async (req: Request, res: Response): Promise<void> => {
    const bodyResult = ProviderSchemas.HttpGenerator.Request.safeParse(req.body);
    if (!bodyResult.success) {
      replyValidationError(res, bodyResult.error);
      return;
    }
    const { requestExample, responseExample } = bodyResult.data;

    if (neverGenerateRemote()) {
      replyError(res, 400, 'Requires remote generation be enabled.');
      return;
    }

    // Strip any trailing slash so we never produce `//api/v1/...`.
    const HOST = (
      cloudConfig.isEnabled()
        ? cloudConfig.getApiHost()
        : getEnvString('PROMPTFOO_CLOUD_API_URL', 'https://api.promptfoo.app')
    ).replace(/\/+$/, '');

    // The fetch layer injects the cloud bearer token for the configured cloud origin
    // (incl. on-prem) and won't override a header we set here, so attaching it
    // explicitly keeps this request authenticated regardless of fetch-layer changes.
    const apiKey = cloudConfig.isEnabled() ? cloudConfig.getApiKey() : undefined;

    try {
      logger.debug('[POST /providers/http-generator] Calling HTTP provider generator API', {
        requestExamplePreview: requestExample?.substring(0, 200),
        hasResponseExample: !!responseExample,
      });

      const response = await fetchWithProxy(`${HOST}/api/v1/http-provider-generator`, {
        method: 'POST',
        headers: {
          'Content-Type': 'application/json',
          ...(apiKey ? { Authorization: `Bearer ${apiKey}` } : {}),
        },
        body: JSON.stringify({
          requestExample,
          responseExample,
        }),
      });

      if (!response.ok) {
        const errorText = await response.text();
        logger.error('[POST /providers/http-generator] Error from cloud API', {
          status: response.status,
          errorText,
        });
        replyError(res, response.status, `HTTP error! status: ${response.status}`);
        return;
      }

      const data = await response.json();
      logger.debug('[POST /providers/http-generator] Successfully generated config');
      res.status(200).json(ProviderSchemas.HttpGenerator.Response.parse(data));
    } catch (error) {
      logger.error('[POST /providers/http-generator] Error calling HTTP provider generator', {
        error,
      });
      sendError(res, 500, 'Failed to generate HTTP configuration');
    }
  },
);

// Test request transform endpoint
providersRouter.post(
  ApiRoutes.Providers.TestRequestTransform.routerPath,
  async (req: Request, res: Response): Promise<void> => {
    const bodyResult = ProviderSchemas.TestRequestTransform.Request.safeParse(req.body);
    if (!bodyResult.success) {
      replyValidationError(res, bodyResult.error, { success: false });
      return;
    }
    const { transformCode, prompt } = bodyResult.data;

    try {
      // Treat empty string as undefined to show base behavior
      const normalizedTransformCode =
        transformCode && transformCode.trim() ? transformCode : undefined;

      // Use the actual HTTP provider's transform function
      const transformFn = await createTransformRequest(normalizedTransformCode);
      const result = await transformFn(
        prompt,
        {},
        { prompt: { raw: prompt, label: prompt }, vars: {} },
      );

      // Check if result is completely empty (no value at all)
      if (result === null || result === undefined) {
        res.json(
          ProviderSchemas.TestRequestTransform.Response.parse({
            success: false,
            error:
              'Transform returned null or undefined. Check your transform function. Did you forget to `return` the result?',
          }),
        );
        return;
      }

      res.json(ProviderSchemas.TestRequestTransform.Response.parse({ success: true, result }));
    } catch (error) {
      const errorMessage = error instanceof Error ? error.message : String(error);
      logger.error('[POST /providers/test-request-transform] Error', {
        error,
      });
      res.status(200).json(
        ProviderSchemas.TestRequestTransform.Response.parse({
          success: false,
          error: errorMessage,
        }),
      );
    }
  },
);

// Test response transform endpoint
providersRouter.post(
  ApiRoutes.Providers.TestResponseTransform.routerPath,
  async (req: Request, res: Response): Promise<void> => {
    const bodyResult = ProviderSchemas.TestResponseTransform.Request.safeParse(req.body);
    if (!bodyResult.success) {
      replyValidationError(res, bodyResult.error, { success: false });
      return;
    }
    const { transformCode, response: responseText } = bodyResult.data;

    try {
      // Treat empty string as undefined to show base behavior
      const normalizedTransformCode =
        transformCode && transformCode.trim() ? transformCode : undefined;

      // Parse the response as JSON if possible
      let jsonData;
      try {
        jsonData = JSON.parse(responseText);
      } catch {
        jsonData = null;
      }

      // Use the actual HTTP provider's transform function
      const transformFn = await createTransformResponse(normalizedTransformCode);
      const result = transformFn(jsonData, responseText);

      // The result is always a ProviderResponse object with an 'output' field
      const output = result?.output ?? result?.raw ?? result;

      if (output === null || output === undefined || output === '') {
        res.json(
          ProviderSchemas.TestResponseTransform.Response.parse({
            success: false,
            error:
              'Transform returned empty result. Ensure that your sample response is correct, and check your extraction path or transform function are returning a valid result.',
            result: JSON.stringify(output),
          }),
        );
        return;
      }

      res.json(
        ProviderSchemas.TestResponseTransform.Response.parse({
          success: true,
          result: output,
        }),
      );
    } catch (error) {
      const errorMessage = error instanceof Error ? error.message : String(error);
      logger.error('[POST /providers/test-response-transform] Error', {
        error,
      });
      res.status(200).json(
        ProviderSchemas.TestResponseTransform.Response.parse({
          success: false,
          error: errorMessage,
        }),
      );
    }
  },
);

// Test multi-turn session functionality
providersRouter.post(
  ApiRoutes.Providers.TestSession.routerPath,
  async (req: Request, res: Response): Promise<void> => {
    const bodyResult = ProviderSchemas.TestSession.Request.safeParse(req.body);
    if (!bodyResult.success) {
      replyValidationError(res, bodyResult.error);
      return;
    }
    const { provider: validatedProvider, sessionConfig, mainInputVariable } = bodyResult.data;

    try {
      const loadedProvider = await loadApiProvider(validatedProvider.id, {
        options: {
          ...validatedProvider,
          config: {
            ...validatedProvider.config,
            maxRetries: 1,
            sessionSource: sessionConfig?.sessionSource || validatedProvider.config?.sessionSource,
            sessionParser: sessionConfig?.sessionParser || validatedProvider.config?.sessionParser,
          },
        },
      });

      // Pass inputs from validatedProvider since loaded provider may not expose config.inputs
      // Check both top-level inputs (from redteam UI) and config.inputs for backwards compatibility
      const result = await testProviderSession({
        provider: loadedProvider,
        sessionConfig,
        inputs: validatedProvider.inputs || validatedProvider.config?.inputs,
        mainInputVariable,
      });

      res.json(ProviderSchemas.TestSession.Response.parse(result));
    } catch (error) {
      sendError(res, 500, 'Failed to test session', error);
    }
  },
);